true-love-base/pyproject.toml
[project]
name = "true-love-base"
version = "0.2.0"
description = "WeChat message processing service based on wxautox4"
requires-python = "==3.12.*"
dependencies = [
    "flask>=3.1.2",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "wxautox4>=40.1.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
true-love-base = "true_love_base.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/true_love_base"]

# $env:UV_PROJECT_ENVIRONMENT="C:\Users\admin\.venvs\true-love-base"
# $env:UV_VENV_PATH=$env:UV_PROJECT_ENVIRONMENT
